backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic-settings==2.5.2
sqlalchemy==2.0.34
psycopg[binary]==3.2.13
reportlab==4.2.5
svglib==1.5.1
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
boto3==1.35.36
pypdfium2==4.30.0
playwright==1.53.0
playwright-stealth==2.0.2
aiohttp>=3.9.0
brotli>=1.1.0
